refactor(web): extract project transfer dialog
