+++ web/src/features/projects/components/TransferProjectButton.tsx
@@ -1,28 +1,6 @@
-/* eslint-disable @repo/no-abstracted-overlay-trigger */
 import { Button } from "@/src/components/ui/button";
-import {
-  Dialog,
-  DialogBody,
-  DialogContent,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-} from "@/src/components/ui/dialog";
-import {
-  Form,
-  FormControl,
-  FormDescription,
-  FormField,
-  FormItem,
-  FormLabel,
-  FormMessage,
-} from "@/src/components/ui/form";
-import { Input } from "@/src/components/ui/input";
+import { TransferProjectDialog } from "@/src/features/projects/components/TransferProjectDialog";
 import { api } from "@/src/utils/api";
-import * as z from "zod";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { usePostHogClientCapture } from "@/src/features/posthog-analytics/usePostHogClientCapture";
 import {
   hasOrganizationAccess,
@@ -30,18 +8,11 @@ import {
 } from "@/src/features/rbac/utils/checkOrganizationAccess";
 import { useQueryProject } from "@/src/features/projects/hooks";
 import { useSession } from "next-auth/react";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/src/components/ui/select";
-import { Alert, AlertDescription, AlertTitle } from "@/src/components/ui/alert";
-import { TriangleAlert } from "lucide-react";
 import { showSuccessToast } from "@/src/features/notifications/showSuccessToast";
+import { useState } from "react";
 
 export function TransferProjectButton() {
+  const [open, setOpen] = useState(false);
   const capture = usePostHogClientCapture();
   const session = useSession();
   const { project, organization } = useQueryProject();
@@ -58,16 +29,6 @@ export function TransferProjectButton() {
         scope: "projects:transfer_org",
       }),
     ) ?? [];
-  const confirmMessage = (organization?.name + "/" + project?.name)
-    .replaceAll(" ", "-")
-    .toLowerCase();
-
-  const formSchema = z.object({
-    name: z.string().includes(confirmMessage, {
-      message: `Please confirm with "${confirmMessage}"`,
-    }),
-    projectId: z.string(),
-  });
 
   const transferProject = api.projects.transfer.useMutation({
     onSuccess: async () => {
@@ -84,121 +45,37 @@ export function TransferProjectButton() {
     },
   });
 
-  const form = useForm({
-    resolver: zodResolver(formSchema),
-    defaultValues: {
-      name: "",
-      projectId: "",
-    },
-  });
-
-  const onSubmit = (values: z.infer<typeof formSchema>) => {
+  const onConfirm = (organizationId: string) => {
     if (!project) return;
     capture("project_settings:project_delete");
     transferProject.mutate({
       projectId: project.id,
-      targetOrgId: values.projectId,
+      targetOrgId: organizationId,
     });
   };
 
   return (
-    <Dialog>
-      <DialogTrigger asChild>
-        <Button variant="destructive-secondary" disabled={!hasAccess}>
-          Transfer Project
-        </Button>
-      </DialogTrigger>
-      <DialogContent className="sm:max-w-[425px]">
-        <DialogHeader>
-          <DialogTitle className="text-lg font-bold">
-            Transfer Project
-          </DialogTitle>
-          <Alert className="mt-2">
-            <TriangleAlert className="h-4 w-4" />
-            <AlertTitle>Warning</AlertTitle>
-            <AlertDescription>
-              Transferring the project will move it to a different organization:
-              <ul className="list-disc pl-4">
-                <li>
-                  Members who are not part of the new organization will lose
-                  access.
-                </li>
-                <li>
-                  The project remains fully operational as API keys, settings,
-                  and data will remain unchanged. All features (e.g. tracing,
-                  prompt management) will continue to work without interruption.
-                </li>
-              </ul>
-            </AlertDescription>
-          </Alert>
-        </DialogHeader>
-        <Form {...form}>
-          <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-            <DialogBody>
-              <FormField
-                control={form.control}
-                name="projectId"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Select New Organization</FormLabel>
-                    <FormControl>
-                      <Select
-                        onValueChange={field.onChange}
-                        value={field.value}
-                        disabled={transferProject.isPending}
-                      >
-                        <SelectTrigger>
-                          <SelectValue placeholder="Select organization" />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {organizationsToTransferTo
-                            .filter((org) => org.id !== organization?.id)
-                            .map((org) => (
-                              <SelectItem key={org.id} value={org.id}>
-                                {org.name}
-                              </SelectItem>
-                            ))}
-                        </SelectContent>
-                      </Select>
-                    </FormControl>
-                    <FormDescription>
-                      Transfer this project to another organization where you
-                      have the ability to create projects.
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-              <FormField
-                control={form.control}
-                name="name"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Confirm</FormLabel>
-                    <FormControl>
-                      <Input placeholder={confirmMessage} {...field} />
-                    </FormControl>
-                    <FormDescription>
-                      {`To confirm, type "${confirmMessage}" in the input box `}
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            </DialogBody>
-            <DialogFooter>
-              <Button
-                type="submit"
-                variant="destructive"
-                loading={transferProject.isPending}
-                className="w-full"
-              >
-                Transfer project
-              </Button>
-            </DialogFooter>
-          </form>
-        </Form>
-      </DialogContent>
-    </Dialog>
+    <>
+      <Button
+        variant="destructive-secondary"
+        disabled={!hasAccess}
+        onClick={() => setOpen(true)}
+      >
+        Transfer Project
+      </Button>
+      {project && organization ? (
+        <TransferProjectDialog
+          open={open}
+          onOpenChange={setOpen}
+          projectName={project.name}
+          organizationName={organization.name}
+          organizations={organizationsToTransferTo.filter(
+            (org) => org.id !== organization.id,
+          )}
+          isPending={transferProject.isPending}
+          onConfirm={onConfirm}
+        />
+      ) : null}
+    </>
   );
 }
